pyproject.toml
[tool.poetry]
name = "todobot"
version = "0.1.0"
description = "Telegram бот для управления списком задач"
authors = ["Dmitrii <lagodnov@gmail.com>"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
aiogram = "^3.5.0"
python-dotenv = "^1.0.1"
sqlalchemy = "^2.0.29"
aiosqlite = "^0.20.0"
asyncpg = "^0.29.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
